feat(schemas,cloud): align tenant types
