+++ packages/console/src/contexts/TenantsProvider.tsx
@@ -24,7 +24,7 @@ type Tenants = {
 const { tenantId, indicator } = defaultManagementApi.resource;
 const initialTenants = conditional(
   !isCloud && [
-    { id: tenantId, name: `tenant_${tenantId}`, tag: `${TenantTag.Development}`, indicator }, // Make `tag` value to be string type.
+    { id: tenantId, name: `tenant_${tenantId}`, tag: TenantTag.Development, indicator }, // Make `tag` value to be string type.
   ]
 );
