+++ packages/schemas/src/types/index.ts
@@ -11,7 +11,6 @@ export * from './role.js';
 export * from './verification-code.js';
 export * from './application.js';
 export * from './system.js';
-export * from './tenant.js';
 export * from './user-assets.js';
 export * from './hook.js';
 export * from './service-log.js';
@@ -19,3 +18,4 @@ export * from './theme.js';
 export * from './cookie.js';
 export * from './dashboard.js';
 export * from './domain.js';
+export * from './tenant/index.js';
