+++ packages/schemas/src/types/tenant/tag.ts
@@ -1,0 +1,5 @@
+export enum TenantTag {
+  Development = 'development',
+  Staging = 'staging',
+  Production = 'production',
+}
